backend/requirements.txt
Django
environs
psycopg2
pytest
pytest-django
pylint-django
djangorestframework
markdown      
django-filter 